+++ packages/rspack/src/ModuleFederationPlugin.ts
@@ -373,3 +373,5 @@ export class ModuleFederationPlugin implements RspackPluginInstance {
 }
 
 export const GetPublicPathPlugin = RemoteEntryPlugin;
+
+export { TreeShakingSharedPlugin } from './TreeShakingSharedPlugin';